+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Translated brands
+
 ## [2.3.0] - 2024-05-20
 
 ## [2.2.0] - 2024-05-14